renames files to avoid confusiong around the word uploading, which is not descriptive in this case
