+++ packages/desktop/src/renderer/components/Channel/Channel.tsx
@@ -10,7 +10,7 @@ import ChannelComponent, { ChannelComponentProps } from './ChannelComponent'
 
 import { useModal } from '../../containers/hooks'
 import { ModalName } from '../../sagas/modals/modals.types'
-import { UploadFilesPreviewsProps } from './File/UploadingPreview'
+import { UploadFilesPreviewsProps } from './File/FilePreview'
 
 import { getFilesData } from '@quiet/common'
