+++ packages/desktop/src/renderer/components/Channel/ChannelComponent.tsx
@@ -17,7 +17,7 @@ import { ChannelMessage, DownloadStatus, Identity, MessagesDailyGroups, MessageS
 import { useResizeDetector } from 'react-resize-detector'
 import { Dictionary } from '@reduxjs/toolkit'
 
-import UploadFilesPreviewsComponent, { UploadFilesPreviewsProps } from './File/UploadingPreview'
+import UploadFilesPreviewsComponent, { UploadFilesPreviewsProps } from './File/FilePreview'
 
 import { DropZoneComponent } from './DropZone/DropZoneComponent'
