+++ packages/desktop/src/renderer/components/Channel/File/FilePreview.tsx
