+++ packages/desktop/src/renderer/components/Channel/File/Image/ImagePreview.tsx
