+++ packages/desktop/src/renderer/components/Channel/File/ImagePlaceholder/ImagePlaceholder.test.tsx
@@ -2,7 +2,7 @@ import { DownloadStatus, DownloadState } from '@quiet/types'
 import React from 'react'
 
 import { renderComponent } from '../../../../testUtils/renderComponent'
-import UploadedImagePlaceholder from './UploadedImagePlaceholder'
+import UploadedImagePlaceholder from './ImagePlaceholder'
 
 describe('UploadedImagePlaceholder', () => {
   let downloadStatus: DownloadStatus
