+++ packages/desktop/src/renderer/components/Channel/File/ImagePlaceholder/ImagePlaceholder.tsx
