+++ packages/desktop/src/renderer/components/widgets/channels/ChannelInput/ChannelInput.test.tsx
@@ -6,7 +6,7 @@ import { ChannelInputComponent } from './ChannelInput'
 import { INPUT_STATE } from './InputState.enum'
 import { getFactory, identity } from '@quiet/state-manager'
 import { FileContent } from '@quiet/types'
-import UploadFilesPreviewsComponent from '../../../Channel/File/UploadingPreview'
+import UploadFilesPreviewsComponent from '../../../Channel/File/FilePreview'
 
 describe('ChannelInput', () => {
   it('renders component input available ', () => {
